Online cmd, update on reload, move formatting to UserManager
