+++ src/main/java/github/vaporrrr/btediscordaddon/commands/discord/Setup.java
@@ -10,13 +10,12 @@ public class Setup extends DiscordCommand {
     public void execute(BTEDiscordAddon bteDiscordAddon, DiscordGuildMessageReceivedEvent event, String[] args) {
         EmbedBuilder embed = new EmbedBuilder();
         embed.setDescription("Minecraft Server Status will be edited into this message. Type " + bteDiscordAddon.getConfig().getString("DiscordCommandsPrefix") + getName() + " again to change the location.");
-        embed.setFooter("Editing config...");
+        embed.setFooter("Updating...");
         event.getChannel().sendMessage(embed.build()).queue((m -> {
             bteDiscordAddon.getConfig().set("ServerStatus.ChannelID", m.getTextChannel().getId());
             bteDiscordAddon.getConfig().set("ServerStatus.MessageID", m.getId());
             bteDiscordAddon.saveConfig();
-            embed.setFooter("Success editing config!");
-            m.editMessage(embed.build()).queue();
+            bteDiscordAddon.getServerStatus().update();
         }));
     }
